Switch to using numpy float datatype (#131)
* Start change of processing to float16

* Update flip node to use numpy rather than oiio

* Update rotate node to use numpy rather than oiio

* Allow for disabling caching of glsl shaders

* Implementation of float16 processing

* Re-enable caches

* Re-implement Mix node to use a glsl shader

* Small additions

* Update to use ``np.float32``

+++ src/nodes/corenodes/input/checkered_image_node.py
@@ -77,7 +77,7 @@ class CheckeredImageNode(api.Node):
 
         render_image = api.RenderImage()
 
-        buf = oiio.ImageBuf(oiio.ImageSpec(1200, 1200, 4, oiio.UINT16))
+        buf = oiio.ImageBuf(oiio.ImageSpec(1200, 1200, 4, oiio.FLOAT))
         oiio.ImageBufAlgo.checker(buf, checker_size, checker_size, 1, 
                                   color1=(0.1,0.1,0.1,255.0), color2=(0.4,0.4,0.4,255.0),
                                   xoffset=offset[0], yoffset=offset[1],
